requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
pandas==2.1.3
openpyxl==3.1.2
gunicorn==21.2.0
